fixed mat's stuff

+++ true_tests/strings/test_string_builder_edge.c
@@ -1,7 +1,7 @@
 // File: true_tests/strings/test_string_builder_edge.c - Edge case tests for string builder
 
-#include "../../include/tests.h"
-#include "../../include/Daedalus.h"
+#include "tests.h"
+#include "Daedalus.h"
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
@@ -289,7 +289,7 @@ int test_string_builder_memory_stress(void)
     for (int i = 0; i < 10; i++) {
         // Grow
         fill_builder_with_content(sb, 100 * (i + 1));
-        TEST_ASSERT(d_StringLen(sb) == 100 * (i + 1), "Should grow correctly in iteration");
+        TEST_ASSERT(d_StringLen(sb) == (size_t)100 * (i + 1), "Should grow correctly in iteration");
 
         // Shrink
         d_StringTruncate(sb, 50);
